extract some common types
