+++ packages/spectron/lib/application.ts
@@ -1,52 +1,11 @@
 import Electron from 'electron';
 import { WaitUntilOptions } from 'webdriverio';
-import { LooseObject } from '@common/types';
+import { LooseObject, SpectronClient } from '@common/types';
 import { ApiName, ApiNames, createApi } from './api';
 /* global browser */
 
-// interface LooseObject {
-//   [key: string]: any;
-// }
-
 function delay(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-declare namespace WebdriverIO {
-  interface BrowserObject {}
-  interface Element {}
-}
-
-export interface SpectronClient extends WebdriverIO.BrowserObject {
-  [key: string]: any;
-  $(): Promise<WebdriverIO.Element>;
-
-  addCommand(commandName: string, executeApiCall: Function): Promise<void>;
-  /**
-   * Wait until the window is no longer loading.
-   * Takes an optional timeout in milliseconds that defaults to 5000.
-   */
-  waitUntilWindowLoaded(timeout?: number): Promise<void>;
-
-  /**
-   * Wait until the element matching the given selector contains the given text.
-   * Takes an optional timeout in milliseconds that defaults to 5000.
-   */
-  waitUntilTextExists(selector: string, text: string, timeout?: number): Promise<void>;
-
-  /**
-   * Gets the number of open windows. <webview> tags are also counted as separate windows.
-   */
-  getWindowCount(): Promise<number>;
-  /**
-   * Focus a window using its index from the windowHandles() array.
-   * <webview> tags can also be focused as a separate window.
-   */
-  windowByIndex(index: number): Promise<void>;
-  /**
-   * Get the selected text in the current window.
-   */
-  getSelectedText(): Promise<string>;
 }
 
 export interface SpectronWindow extends Electron.BrowserWindow {
